+++ src/robinhood/rbh_du.c
@@ -127,8 +127,8 @@ static lmgr_filter_t    parent_filter; /* same as entry_filter + condition on pa
 static bool_node_t      match_expr;
 static int              is_expr = 0; /* is it set? */
 
-static int disp_mask = ATTR_MASK_type | ATTR_MASK_blocks | ATTR_MASK_size;
-static int query_mask = 0;
+static uint64_t disp_mask = ATTR_MASK_type | ATTR_MASK_blocks | ATTR_MASK_size;
+static uint64_t query_mask = 0;
 
 typedef struct stats_du_t
 {
